Feat:updated initialcardstrick
